requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
blas=1.0=mkl
brotli-python=1.0.9=py39h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.7.2=h06a4308_0
certifi=2024.7.4=py39h06a4308_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
contourpy=1.2.1=pypi_0
cuda=11.6.1=0
cuda-cccl=11.6.55=hf6102b2_0
cuda-command-line-tools=11.6.2=0
cuda-compiler=11.6.2=0
cuda-cudart=11.6.55=he381448_0
cuda-cudart-dev=11.6.55=h42ad0f4_0
cuda-cuobjdump=11.6.124=h2eeebcb_0
cuda-cupti=11.6.124=h86345e5_0
cuda-cuxxfilt=11.6.124=hecbf4f6_0
cuda-driver-dev=11.6.55=0
cuda-gdb=12.5.82=0
cuda-libraries=11.6.1=0
cuda-libraries-dev=11.6.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=12.5.82=0
cuda-nsight-compute=12.4.1=0
cuda-nvcc=11.6.124=hbba6d2d_0
cuda-nvdisasm=12.5.39=0
cuda-nvml-dev=11.6.55=haa9ef22_0
cuda-nvprof=12.5.82=0
cuda-nvprune=11.6.124=he22ec0a_0
cuda-nvrtc=11.6.124=h020bade_0
cuda-nvrtc-dev=11.6.124=h249d397_0
cuda-nvtx=11.6.124=h0630a44_0
cuda-nvvp=12.5.82=0
cuda-runtime=11.6.1=0
cuda-samples=11.6.101=h8efea70_0
cuda-sanitizer-api=12.5.81=0
cuda-toolkit=11.6.1=0
cuda-tools=11.6.1=0
cuda-version=12.5=3
cuda-visual-tools=11.6.1=0
cycler=0.12.1=pypi_0
dbus=1.13.18=hb2f20db_0
expat=2.6.2=h6a678d5_0
ffmpeg=4.3=hf484d3e_0
fontconfig=2.14.1=h4c34cd2_2
fonttools=4.53.1=pypi_0
freetype=2.12.1=h4a9f257_0
gds-tools=1.10.1.7=0
glib=2.78.4=h6a678d5_0
glib-tools=2.78.4=h6a678d5_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpcio=1.65.2=pypi_0
icu=73.1=h6a678d5_0
idna=3.7=py39h06a4308_0
importlib-metadata=8.2.0=pypi_0
importlib-resources=6.4.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
jpeg=9e=h5eee18b_2
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.9.2.110=h5e84587_0
libcublas-dev=11.9.2.110=h5c901ab_0
libcufft=10.7.1.112=hf425ae0_0
libcufft-dev=10.7.1.112=ha5ce4c0_0
libcufile=1.10.1.7=0
libcufile-dev=1.10.1.7=0
libcurand=10.3.6.82=0
libcurand-dev=10.3.6.82=0
libcusolver=11.3.4.124=h33c3c4e_0
libcusparse=11.7.2.124=h7538f96_0
libcusparse-dev=11.7.2.124=hbbe9722_0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libglib=2.78.4=hdc74915_0
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libnpp=11.6.3.124=hd2722f0_0
libnpp-dev=11.6.3.124=h3c42840_0
libnvjpeg=11.6.2.124=hd473ad6_0
libnvjpeg-dev=11.6.2.124=hb5906b9_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=h5eee18b_1
libxml2=2.10.4=hfdd30dd_2
lz4-c=1.9.4=h6a678d5_1
markdown=3.6=pypi_0
markupsafe=2.1.5=pypi_0
matplotlib=3.9.1=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py39h5eee18b_1
mkl_fft=1.3.8=py39h5eee18b_0
mkl_random=1.2.4=py39hdb19cb5_0
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
nsight-compute=2024.2.1.2=2
nspr=4.35=h6a678d5_0
nss=3.89.1=h6a678d5_0
numpy=1.26.4=py39h5f9d8c6_0
numpy-base=1.26.4=py39hb5e798b_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h9ca470c_2
openssl=3.0.14=h5eee18b_0
packaging=24.1=pypi_0
pandas=2.2.2=pypi_0
pcre2=10.42=hebb0a14_1
pillow=10.4.0=py39h5eee18b_0
pip=24.0=py39h06a4308_0
protobuf=4.25.4=pypi_0
pyparsing=3.1.2=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.19=h955ad1f_1
python-dateutil=2.9.0.post0=pypi_0
pytorch=1.13.1=py3.9_cuda11.6_cudnn8.3.2_0
pytorch-cuda=11.6=h867d48c_1
pytorch-mutex=1.0=cuda
pytz=2024.1=pypi_0
readline=8.2=h5eee18b_0
requests=2.32.3=py39h06a4308_0
scipy=1.13.1=pypi_0
setuptools=69.5.1=py39h06a4308_0
six=1.16.0=pypi_0
sqlite=3.45.3=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.17.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
thop=0.1.1-2209072238=pypi_0
tk=8.6.14=h39e8969_0
torchvision=0.14.1=py39_cu116
tqdm=4.66.4=pypi_0
typing_extensions=4.11.0=py39h06a4308_0
tzdata=2024.1=pypi_0
urllib3=2.2.2=py39h06a4308_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py39h06a4308_0
xz=5.4.6=h5eee18b_1
zipp=3.19.2=pypi_0
zlib=1.2.13=h5eee18b_1
zstd=1.5.5=hc292b87_2
